feature/recuperar_senha verificacao de existencia de usuario e remocao do servico de localizacao de senhas

+++ src/main/java/br/com/appetitegourmet/api/spring/login/controllers/AuthController.java
@@ -167,6 +167,10 @@ public class AuthController {
 	String password;
 	Boolean falhou = false;
 	
+	if(!userRepository.existsByEmail(loginRequest.getUsername())) {
+		return ResponseEntity.badRequest().body(new MessageResponse("Usuario inexistente!"));
+	}
+	
 	password = gs.GeradorSenha();
 	userService.removeLink(loginRequest.getUsername());
 	userService.salvarHashSenha(loginRequest.getUsername(), password);
